+++ src/in/mathutil.h
@@ -11,4 +11,4 @@ Complex comp_add(Complex a, Complex b);
 Complex comp_mul(Complex a, Complex b);
 
 // Returns the position of given number on given axis and its boundaries on the complex plane
-double map_point_to_complex(double val, double in_max, double in_min, double out_max, double out_min);
+long double map_point_to_complex(double val, double in_max, double in_min, double out_max, double out_min);
